// frontend/src/api/models/LearnBlock.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LearnBlock
 */
export interface LearnBlock {
    /**
     * 
     * @type {number}
     * @memberof LearnBlock
     */
    position: number;
    /**
     * 
     * @type {string}
     * @memberof LearnBlock
     */
    content: string;
    /**
     * 
     * @type {number}
     * @memberof LearnBlock
     */
    learnId: number;
    /**
     * 
     * @type {number}
     * @memberof LearnBlock
     */
    moduleId: number;
}

/**
 * Check if a given object implements the LearnBlock interface.
 */
export function instanceOfLearnBlock(value: object): value is LearnBlock {
    if (!('position' in value) || value['position'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('learnId' in value) || value['learnId'] === undefined) return false;
    if (!('moduleId' in value) || value['moduleId'] === undefined) return false;
    return true;
}

export function LearnBlockFromJSON(json: any): LearnBlock {
    return LearnBlockFromJSONTyped(json, false);
}

export function LearnBlockFromJSONTyped(json: any, ignoreDiscriminator: boolean): LearnBlock {
    if (json == null) {
        return json;
    }
    return {
        
        'position': json['position'],
        'content': json['content'],
        'learnId': json['learn_id'],
        'moduleId': json['module_id'],
    };
}

export function LearnBlockToJSON(json: any): LearnBlock {
    return LearnBlockToJSONTyped(json, false);
}

export function LearnBlockToJSONTyped(value?: LearnBlock | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'position': value['position'],
        'content': value['content'],
        'learn_id': value['learnId'],
        'module_id': value['moduleId'],
    };
}